refactor(composer): move DSL helpers to factories.py for better architectural purity
